todonotes/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.2
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.13.1
factory-boy==3.2.1
Faker==13.3.0
idna==3.3
importlib-metadata==4.11.2
Markdown==3.3.6
optional-django==0.3.0
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.8
zipp==3.7.0
